backend/pyproject.toml
[project]
name = "fertiscan-backend"
version = "0.1.0"
description = "FertiScan Backend API with FastAPI and SQLAlchemy"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "fastapi[standard]",
    "pydantic-settings",
    "sqlmodel",
    "alembic",
    "bcrypt",
    "pyjwt",
    "jinja2",
    "aiosmtplib",
    "tenacity",
    "psycopg[binary]",
    "fastapi-pagination>=0.15.5",
    "aioboto3",
    "openai",
    "instructor",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xdist>=3.5.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.24.0",
    "mypy>=1.18.0",
    "ruff>=0.2.0",
    "coverage>=7.4.0",
    "pre-commit>=3.6.0",
    "types-jinja2>=2.11.0",
    "types-PyJWT>=1.7.1",
    "boto3-stubs[s3]>=1.42.25",
    "aiomoto[s3]>=0.1.0",
]
sbom = [
    "cyclonedx-bom>=7.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
exclude = ["alembic", "app/data"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic", "app/data"]
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.pytest.ini_options]
addopts = "-n auto --dist=loadfile"
asyncio_mode = "auto"
